+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Unreleased
   compression headers
 - Adjusted symbol names reported for BPF programs to contain `bpf_prog_`
   prefix and program tag
+- Made `inspect::SymInfo` type non-exhaustive
 - Changed `CodeInfo::to_owned` to `into_owned`
 - Bumped minimum supported Rust version to `1.75`